LocalDataSourceImplTest 작업중

+++ data/src/main/java/com/teamtuna/emotionaldiary/repository/EmotionRepositoryImpl.kt
@@ -4,7 +4,7 @@ import com.teamtuna.emotionaldiary.datasource.LocalDataSource
 import com.teamtuna.emotionaldiary.entity.Emotion
 
 class EmotionRepositoryImpl(private val localDataSource: LocalDataSource) : EmotionRepository {
-    override fun add(emotion: Emotion, reason: String): Int {
+    override fun add(emotion: Emotion, reason: String): Long {
         return localDataSource.add(emotion, reason)
     }
 }
