+++ domain/src/main/java/com/teamtuna/emotionaldiary/repository/EmotionRepository.kt
@@ -3,5 +3,5 @@ package com.teamtuna.emotionaldiary.repository
 import com.teamtuna.emotionaldiary.entity.Emotion
 
 interface EmotionRepository {
-    fun add(emotion: Emotion, reason: String): Int
+    fun add(emotion: Emotion, reason: String): Long
 }
